Restructure BookDetail.java
